src/subject_space/requirements.txt
azure-servicebus==7.12.2
fastapi==0.112.0
python-dotenv==1.0.1
azure-cosmos==4.7.0
